framrates admin #1138
https://github.com/owncast/owncast/issues/1138#

This is a technically correct and low to no risk making the low end default 25.

now has 25 default, 30, 50, 60 and outrageous. This is a good balance as the standard in NA and Japan and such and such is still 30/60 and EU etc is 25/50.  So I've just piggybacked the same description for 50 and 60 options and tacked an EU and NA on them.

I don't feel this is overwhelming or confusing  as it is in Advanced, and it's is legit nice for source for the world users...

Pardon my ignorance bu I cannot find where the default stream output name is coming from, tried reinstalling owncast on top of my updated admin but it still sticks a defualt NAME of "1.2Mpbs@24fps" in the admin UI even though create and edit work great with the above changes.


This is a technically correct and low to no risk making the low end default 25.

The paradox is I am having heck of a time locating how the initial default stream output is setup anywhere in the owncast deployment. It is not a big deal as anyone who doesnt care will have the default 24fps anyone who does care will go to advanced and make their own variant and delete the default. 

now has 25 default, 30, 50, 60 and outrageous. This is a good balance as the standard in NA and Japan and such and such is still 30/60 and EU etc is 25/50.  So I've just piggybacked the same description for 50 and 60 options and tacked an EU and NA on them.

tested streaming with all variants individually and stacked the with no issues.